split synctexjs.ts into two files
